requirements.txt
aiohttp==3.7.3
APScheduler==3.7.0
async-timeout==3.0.1
attrs==19.3.0
azure-core==1.12.0b1
azure-eventhub==5.3.0
azure-eventhub-checkpointstoreblob-aio==1.1.2
azure-storage-blob==12.8.0b1
Brotli==1.0.9
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
Click==7.0
colorhash==1.0.3
configparser==5.0.1
cryptography==3.4.6
cycler==0.10.0
dash==1.19.0
dash-core-components==1.15.0
dash-daq==0.5.0
dash-html-components==1.1.2
dash-renderer==1.9.0
dash-table==4.11.2
Flask==1.1.1
Flask-Compress==1.9.0
Flask-Cors==3.0.8
Flask-MonitoringDashboard==3.0.6
future==0.18.2
idna==2.10
idna-ssl==1.1.0
importlib-metadata==1.4.0
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.0
joblib==0.14.1
jsonschema==3.2.0
kiwisolver==1.1.0
kneed==0.5.1
MarkupSafe==1.1.1
matplotlib==3.1.2
more-itertools==8.1.0
msrest==0.6.21
multidict==5.1.0
numpy==1.18.1
oauthlib==3.1.0
pandas==0.25.3
plotly==4.14.3
psutil==5.8.0
pycparser==2.20
pymongo==3.11.3
pyparsing==2.4.6
pyrsistent==0.15.7
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
regexp==0.1
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
scikit-learn==0.22.1
scipy==1.5.4
six==1.14.0
sklearn==0.0
SQLAlchemy==1.3.22
typing-extensions==3.7.4.3
tzlocal==2.1
uamqp==1.2.14
urllib3==1.26.3
Werkzeug==0.16.1
wincertstore==0.2
xgboost==0.90
yarl==1.6.3
zipp==2.0.1
